+++ app/main.js
@@ -1,5 +1,5 @@
 // process.on('uncaughtException', (...args) => console.error(...args))
-const {app, BrowserWindow, Tray, Menu, ipcMain, shell, dialog} = require('electron')
+const {app, BrowserWindow, Tray, Menu, ipcMain, shell, dialog, globalShortcut} = require('electron')
 const path = require('path')
 const AppSettings = require('./utils/settings')
 const defaultSettings = require('./utils/defaultSettings')
@@ -101,7 +101,9 @@ function startMicrobreak () {
     console.log('microbreak already running')
     return
   }
-
+  globalShortcut.register('CommandOrControl+X', () => {
+    finishMicrobreak(false)
+  })
   const modalPath = path.join('file://', __dirname, 'microbreak.html')
   microbreakWin = new BrowserWindow({
     x: displaysX(),
@@ -110,6 +112,8 @@ function startMicrobreak () {
     show: false,
     fullscreen: settings.get('fullscreen'),
     backgroundColor: settings.get('mainColor'),
+    skipTaskbar: true,
+    focusable: false,
     title: 'stretchly'
   })
   microbreakWin.loadURL(modalPath)
@@ -132,6 +136,9 @@ function startBreak () {
     console.log('break already running')
     return
   }
+  globalShortcut.register('CommandOrControl+X', () => {
+    finishBreak(false)
+  })
   const modalPath = path.join('file://', __dirname, 'break.html')
   breakWin = new BrowserWindow({
     x: displaysX(),
@@ -140,6 +147,8 @@ function startBreak () {
     show: false,
     fullscreen: settings.get('fullscreen'),
     backgroundColor: settings.get('mainColor'),
+    skipTaskbar: true,
+    focusable: false,
     title: 'stretchly'
   })
   breakWin.loadURL(modalPath)
@@ -154,11 +163,16 @@ function startBreak () {
 }
 
 function finishMicrobreak (shouldPlaySound = true) {
+  globalShortcut.unregister('CommandOrControl+X')
   clearTimeout(finishMicrobreakTimer)
   if (shouldPlaySound) {
     processWin.webContents.send('playSound', settings.get('audio'))
   }
   if (microbreakWin) {
+    if (process.platform === 'darwin') {
+      // get focus on the last app
+      Menu.sendActionToFirstResponder('hide:')
+    }
     microbreakWin.close()
     microbreakWin = null
     breakPlanner.nextBreak.plan()
@@ -166,11 +180,16 @@ function finishMicrobreak (shouldPlaySound = true) {
 }
 
 function finishBreak (shouldPlaySound = true) {
+  globalShortcut.unregister('CommandOrControl+X')
   clearTimeout(finishBreakTimer)
   if (shouldPlaySound) {
     processWin.webContents.send('playSound', settings.get('audio'))
   }
   if (breakWin) {
+    if (process.platform === 'darwin') {
+      // get focus on the last app
+      Menu.sendActionToFirstResponder('hide:')
+    }
     breakWin.close()
     breakWin = null
     breakPlanner.nextBreak.plan()